Feature/68 user module

+++ api/v1/modules/user/controllers/UserController.ts
@@ -1,6 +1,7 @@
 // eslint-disable-next-line @typescript-eslint/await-thenable
-import { type Request, type Response } from 'express';
+import e, { type Request, type Response } from 'express';
 import UserService from '../services/UserService';
+import UserMiddleware from '../middlewares/UserMiddleware';
 
 // Default controller functions
 export const getAllUser: any = async (req: Request, res: Response) => {
@@ -16,29 +17,116 @@ export const getAllUser: any = async (req: Request, res: Response) => {
 export const getUserById: any = async (req: Request, res: Response) => {
   try {
     const id = req.params.id;
-    const data = await UserService.getUserById(id);
-    res.status(200).send(data);
+    const token = req.headers.authorization ?? '';
+    const check = await UserMiddleware.checkGetUserById(id, token);
+    if (check === 'User not found') {
+      res.status(404).send(check);
+    } else if (check === 'Unauthorized' || check === 'No token provided') {
+      res.status(401).send(check);
+    } else {
+      const data = await UserService.getUserById(id);
+      res.status(200).send(data);
+    }
   } catch (error: any) {
     console.error(error);
     res.status(500).send(error.message);
   }
 };
 
-export const createUser: any = async (req: Request, res: Response) => {
+export const createCandidate: any = async (req: Request, res: Response) => {
   try {
-    const data = await UserService.createUser(req.body);
-    res.status(200).send(data);
+    const check = await UserMiddleware.checkCreateCandidate(req.body);
+    if (check === 'Missing required fields') {
+      res.status(400).send(check);
+    } else if (check === 'Username already exists' || check === 'User with that email already exists' || check === 'User with that GitHub username already exists') {
+      res.status(409).send(check);
+    } else {
+      const role: string = 'Candidate';
+      const data = await UserService.createUser(req.body, role);
+      res.status(200).send(data);
+    }
   } catch (error: any) {
     console.error(error);
     res.status(500).send(error.message);
   }
 };
 
-export const updateUser: any = async (req: Request, res: Response) => {
+export const createRepresentative: any = async (req: Request, res: Response) => {
+  try {
+    const check = await UserMiddleware.checkCreateRepresentative(req.body);
+    if (check === 'Missing required fields') {
+      res.status(400).send(check);
+    } else if (check === 'Username already exists' || check === 'User with that email already exists') {
+      res.status(409).send(check);
+    } else {
+      const role: string = 'Representative';
+      const data = await UserService.createUser(req.body, role);
+      res.status(200).send(data);
+    }
+  } catch (error: any) {
+    console.error(error);
+    res.status(500).send(error.message);
+  }
+};
+
+export const updateCandidate: any = async (req: Request, res: Response) => {
   try {
     const id = req.params.id;
-    const data = await UserService.updateUser(id, req.body);
-    res.status(200).send(data);
+    const token = req.headers.authorization ?? '';
+    const check = await UserMiddleware.checkUpdateCandidate(id, token, req.body);
+    if (check === 'User not found') {
+      res.status(404).send(check);
+    } else if (check === 'No data to update') {
+      res.status(400).send(check);
+    } else if (check === 'Unauthorized' || check === 'No token provided') {
+      res.status(401).send(check);
+    } else {
+      const role: string = 'Candidate';
+      const data = await UserService.updateUser(id, req.body, role);
+      res.status(200).send(data);
+    }
+  } catch (error: any) {
+    console.error(error);
+    res.status(500).send(error.message);
+  }
+};
+
+export const updateRepresentative: any = async (req: Request, res: Response) => {
+  try {
+    const id = req.params.id;
+    const token = req.headers.authorization ?? '';
+    const check = await UserMiddleware.checkUpdateRepresentative(id, token, req.body);
+    if (check === 'User not found') {
+      res.status(404).send(check);
+    } else if (check === 'No data to update') {
+      res.status(400).send(check);
+    } else if (check === 'Unauthorized' || check === 'No token provided') {
+      res.status(401).send(check);
+    } else {
+      const role: string = 'Representative';
+      const data = await UserService.updateUser(id, req.body, role);
+      res.status(200).send(data);
+    }
+  } catch (error: any) {
+    console.error(error);
+    res.status(500).send(error.message);
+  }
+}
+
+export const loginUser: any = async (req: Request, res: Response) => {
+  try {
+    const token = req.headers.authorization ?? '';
+    const check = await UserMiddleware.checkLoginUser(token, req.body);
+    if (check === 'User not found') {
+      res.status(404).send(check);
+    } else if (check === 'Invalid password') {
+      res.status(401).send(check);
+    } else if (check === 'User already logged in') {
+      res.status(401).send(check);
+    } else {
+      const data = await UserService.loginUser(req.body);
+      res.status(200).send(data);
+    }
   } catch (error: any) {
     console.error(error);
     res.status(500).send(error.message);
@@ -48,7 +136,58 @@ export const updateUser: any = async (req: Request, res: Response) => {
 export const deleteUser: any = async (req: Request, res: Response) => {
   try {
     const id = req.params.id;
-    const data = await UserService.deleteUser(id);
+    const token = req.headers.authorization ?? '';
+    const check = await UserMiddleware.checkDeleteUser(id, token);
+    if (check === 'User not found') {
+      res.status(404).send(check);
+    } else if (check === 'Unauthorized' || check === 'No token provided') {
+      res.status(401).send(check);
+    } else {
+      const data = await UserService.deleteUser(id);
+      res.status(200).send(data);
+    }
+  } catch (error: any) {
+    console.error(error);
+    res.status(500).send(error.message);
+  }
+};
+
+export const getProfessionalExperienceByUserId: any = async (req: Request, res: Response) => {
+  try {
+    const userId = req.params.id;
+    const data = await UserService.getProfessionalExperienceByUserId(userId);
+    res.status(200).send(data);
+  } catch (error: any) {
+    console.error(error);
+    res.status(500).send(error.message);
+  }
+};
+
+export const createProfessionalExperience: any = async (req: Request, res: Response) => {
+  try {
+    const data = await UserService.createProfessionalExperience(req.body);
+    res.status(200).send(data);
+  } catch (error: any) {
+    console.error(error);
+    res.status(500).send(error.message);
+  }
+};
+
+export const updateProfessionalExperience: any = async (req: Request, res: Response) => {
+  try {
+    const id = req.params.id;
+    const data = await UserService.updateProfessionalExperience(id, req.body);
+    res.status(200).send(data);
+  } catch (error: any) {
+    console.error(error);
+    res.status(500).send(error.message);
+  }
+}
+
+export const deleteProfessionalExperience: any = async (req: Request, res: Response) => {
+  try {
+    const id = req.params.id;
+    const data = await UserService.deleteProfessionalExperience(id);
     res.status(200).send(data);
   } catch (error: any) {
     console.error(error);
@@ -58,7 +197,14 @@ export const deleteUser: any = async (req: Request, res: Response) => {
 export default {
   getAllUser,
   getUserById,
-  createUser,
-  updateUser,
-  deleteUser
+  createCandidate,
+  createRepresentative,
+  updateCandidate,
+  updateRepresentative,
+  deleteUser,
+  loginUser,
+  getProfessionalExperienceByUserId,
+  createProfessionalExperience,
+  updateProfessionalExperience,
+  deleteProfessionalExperience
 };
